[FEAT] Design.swift
* 디자인 관련 extension을 별도의 파일로 작성했음.
